refactor: Split async contract-test service to mirror sync layout
